requirements.txt
Flask==3.0.1
Pyrebase4==4.7.1
python-dotenv==1.0.1
Requests==2.31.0
